requirements.txt
alembic==1.3.1
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
idna==2.8
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.4
pycparser==2.19
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
redis==3.3.11
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.8
tornado==6.0.3
urllib3==1.25.7
